Feat - Refactor project page(#38)
