+++ src/component/project/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { useParams } from 'react-router-dom';
 import {
   Banner,
@@ -11,36 +11,14 @@ import {
 import { Body1, Header2, Inner, Section } from '../emotion/GlobalStyle';
 import { useSelectBoxes } from './hook';
 import { ProjectBoxProps } from '../../types/globalType';
-import { useGetVideosInfinityQuery } from '../../store/projectController';
 import { LoadingContainer } from './component';
 
 const Index = () => {
   const { sort } = useParams();
-  const [pageNumber, setPageNumber] = useState(0);
 
-  const { data, isFetching } = useGetVideosInfinityQuery({ size: 6, page: pageNumber });
-  const { sortType, optionType, handleSelectChange } = useSelectBoxes(
+  const { sortType, optionType, handleSelectChange, data, isFetching, pageNumber } = useSelectBoxes(
     sort === 'popular' ? '인기도 순' : '최신 등록순',
   );
-
-  const handleScroll = () => {
-    // 스크롤이 아래로 내려갔을 때의 처리
-    if (
-      window.innerHeight + window.scrollY >= document.body.offsetHeight &&
-      !isFetching &&
-      data?.totalPages !== pageNumber
-    ) {
-      setPageNumber(pageNumber + 1);
-      console.log('도달!');
-    }
-  };
-
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll);
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, [isFetching]);
 
   return (
     <Inner>
